requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic>=2.5.0
pydantic-settings>=2.1.0
torch>=2.2.0
numpy>=1.24.3
librosa==0.10.1
soundfile==0.12.1
scipy==1.11.3
scikit-learn==1.3.2
python-dotenv==1.0.0 